requirements.txt
langchain
langchain-openai
openai
chromadb
python-dotenv
rich
unstructured
tiktoken
beautifulsoup4